Move func2caller into accel-derive
