+++ accel-derive/Cargo.toml
@@ -15,17 +15,11 @@ categories    = []
 proc-macro = true
 
 [dependencies]
-synom = "0.11.3"
-quote = "0.4"
-toml = "0.4.5"
-tempdir = "0.3"
-glob = "0.2.11"
-procedurals = "0.2"
-serde = "1.0"
-serde_derive = "1.0"
-derive_more = "0.10"
+quote = "0.5"
+
+[dependencies.nvptx]
+path = "../nvptx"
 
 [dependencies.syn]
-version = "0.12"
-features = ["full", "fold", "parsing", "printing", "extra-traits"]
-default-features = false
+version = "0.13"
+features = ["full", "extra-traits"]
